Allow tags to be queried even when a field key exists with the same name
A measurement can have a field key on some series, but the field isn't
present on other series in that same measurement. If the series has a
tag with that name, that tag will be returned instead.

A series still cannot have both a field key and a tag key with the same name.

Fixes #4630.